+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PATCH version when backwards compatible bug **fixes** are implemented.
 
 ## [Unreleased]
+### Added
+- Put method
 
 ## [0.1.1] - 2025-04-22
 ### Fixed